# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Svarog - Solana Token Launcher Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  earningsToday: number;
  earningsLast30Days: number;
  bundlesLaunched: number;
  totalBalanceSol: number;
  totalBalanceUsd: number;
  pnl: number;
  activeBots?: number;
  totalWallets?: number;
}

export interface ActivityEvent {
  id: number;
  type: string;
  description: string;
  timestamp: string;
  /** @nullable */
  tokenName?: string | null;
  /** @nullable */
  tokenSymbol?: string | null;
  /** @nullable */
  amount?: number | null;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  walletAddress?: string | null;
}

export interface Wallet {
  id: number;
  publicKey: string;
  ownerAddress: string;
  /** @nullable */
  label: string | null;
  /** @nullable */
  group: string | null;
  /** @nullable */
  balanceSol?: number | null;
  /** @nullable */
  balanceUsd?: number | null;
  isActive?: boolean;
  createdAt: string;
  updatedAt?: string;
}

export interface WalletInput {
  ownerAddress: string;
  label?: string;
  group?: string;
}

export interface WalletImport {
  ownerAddress: string;
  privateKey: string;
  label?: string;
  group?: string;
}

export interface BulkWalletInput {
  ownerAddress: string;
  count: number;
  group?: string;
}

export interface WalletUpdate {
  label?: string;
  group?: string;
  isActive?: boolean;
}

export interface WalletBalance {
  publicKey: string;
  balanceSol: number;
  /** @nullable */
  balanceUsd?: number | null;
}

export interface WalletGroup {
  name: string;
  count: number;
  /** @nullable */
  totalBalanceSol?: number | null;
}

export interface Bundle {
  id: number;
  ownerAddress: string;
  tokenName: string;
  tokenSymbol: string;
  /** @nullable */
  tokenAddress: string | null;
  /** @nullable */
  tokenDescription?: string | null;
  /** @nullable */
  tokenImageUrl?: string | null;
  walletCount: number;
  /** @nullable */
  solPerWallet?: number | null;
  /** @nullable */
  totalSolSpent?: number | null;
  status: string;
  launchType: string;
  /** @nullable */
  performanceUsd?: number | null;
  /** @nullable */
  performanceSol?: number | null;
  /** @nullable */
  txHash?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface BundleInput {
  ownerAddress: string;
  tokenName: string;
  tokenSymbol: string;
  tokenDescription?: string;
  tokenImageUrl?: string;
  walletCount: number;
  solPerWallet?: number;
  walletIds?: number[];
}

export interface VampInput {
  ownerAddress: string;
  sourceTokenAddress: string;
  walletCount: number;
  solPerWallet?: number;
}

export interface BundleStats {
  totalLaunches: number;
  bestLaunchUsd: number;
  bestLaunchSol?: number;
  totalSolSpent: number;
  successRate: number;
  avgWalletsPerBundle?: number;
}

export interface BumpBot {
  id: number;
  ownerAddress: string;
  tokenAddress: string;
  /** @nullable */
  tokenName?: string | null;
  walletCount: number;
  speed: string;
  budgetSol: number;
  /** @nullable */
  solPerWallet?: number | null;
  /** @nullable */
  bumpSize?: number | null;
  /** @nullable */
  bumpsPerHour?: number | null;
  /** @nullable */
  costPerBump?: number | null;
  /** @nullable */
  totalBumps?: number | null;
  /** @nullable */
  bumpsExecuted?: number | null;
  /** @nullable */
  estimatedDurationHours?: number | null;
  status: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  stoppedAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface BotInput {
  ownerAddress: string;
  tokenAddress: string;
  walletCount: number;
  speed: string;
  budgetSol: number;
}

export interface BotEstimateInput {
  walletCount: number;
  speed: string;
  budgetSol: number;
}

export interface BotEstimate {
  bumpSize: number;
  bumpsPerHour: number;
  costPerBump: number;
  totalBumps: number;
  estimatedDurationHours: number;
  solPerWallet: number;
  isValid?: boolean;
  /** @nullable */
  validationMessage?: string | null;
}

export interface TokenMetadata {
  address: string;
  name: string;
  symbol: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  decimals?: number | null;
  /** @nullable */
  supply?: number | null;
}

export interface UserSettings {
  walletAddress: string;
  defaultWalletCount?: number;
  defaultSolPerWallet?: number;
  autoApprove?: boolean;
  darkMode?: boolean;
  notificationsEnabled?: boolean;
  /** @nullable */
  rpcEndpoint?: string | null;
  createdAt?: string;
  updatedAt?: string;
}

export interface SettingsUpdate {
  walletAddress?: string;
  defaultWalletCount?: number;
  defaultSolPerWallet?: number;
  autoApprove?: boolean;
  darkMode?: boolean;
  notificationsEnabled?: boolean;
  rpcEndpoint?: string;
}

export type GetDashboardStatsParams = {
walletAddress?: string;
};

export type GetDashboardActivityParams = {
walletAddress?: string;
limit?: number;
};

export type ListWalletsParams = {
ownerAddress?: string;
group?: string;
};

export type ListWalletGroupsParams = {
ownerAddress?: string;
};

export type ListBundlesParams = {
ownerAddress?: string;
search?: string;
status?: string;
};

export type GetBundleStatsParams = {
ownerAddress?: string;
};

export type ListBotsParams = {
ownerAddress?: string;
status?: string;
};

export type GetTokenMetadataParams = {
ca: string;
};

export type GetSettingsParams = {
walletAddress?: string;
};