# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Svarog - Solana Token Launcher Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bundles:
    get:
      operationId: listBundles
      tags: [bundles]
      summary: List all bundles
      parameters:
        - name: ownerAddress
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of bundles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bundle"
    post:
      operationId: createBundle
      tags: [bundles]
      summary: Create and launch a new token bundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BundleInput"
      responses:
        "201":
          description: Created bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
  /bundles/vamp:
    post:
      operationId: createVampBundle
      tags: [bundles]
      summary: VAMP - copy existing token metadata and relaunch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VampInput"
      responses:
        "201":
          description: Created VAMP bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
  /bundles/stats:
    get:
      operationId: getBundleStats
      tags: [bundles]
      summary: Get bundle performance stats
      parameters:
        - name: ownerAddress
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Bundle stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BundleStats"
  /bundles/{id}:
    get:
      operationId: getBundle
      tags: [bundles]
      summary: Get a specific bundle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bundle"
        "404":
          description: Not found
    delete:
      operationId: deleteBundle
      tags: [bundles]
      summary: Delete a bundle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Bump Bots
  /bots:
    get:
      operationId: listBots
      tags: [bots]
      summary: List all bump bots
      parameters:
        - name: ownerAddress
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BumpBot"
    post:
      operationId: createBot
      tags: [bots]
      summary: Create a new bump bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotInput"
      responses:
        "201":
          description: Created bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BumpBot"
  /bots/{id}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get a specific bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BumpBot"
        "404":
          description: Not found
    delete:
      operationId: deleteBot
      tags: [bots]
      summary: Delete a bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /bots/{id}/start:
    post:
      operationId: startBot
      tags: [bots]
      summary: Start a bump bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BumpBot"
  /bots/{id}/stop:
    post:
      operationId: stopBot
      tags: [bots]
      summary: Stop a bump bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BumpBot"
  /bots/{id}/pause:
    post:
      operationId: pauseBot
      tags: [bots]
      summary: Pause a bump bot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BumpBot"
  /bots/estimate:
    post:
      operationId: estimateBotCost
      tags: [bots]
      summary: Estimate bump bot cost and performance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotEstimateInput"
      responses:
        "200":
          description: Bot cost estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotEstimate"

  # Tokens
...
components:
  schemas:
...
    Bundle:
      type: object
      required: [id, ownerAddress, tokenName, tokenSymbol, tokenAddress, walletCount, status, launchType, createdAt]
      properties:
        id:
          type: integer
        ownerAddress:
          type: string
        tokenName:
          type: string
        tokenSymbol:
          type: string
        tokenAddress:
          type: ["string", "null"]
        tokenDescription:
          type: ["string", "null"]
        tokenImageUrl:
          type: ["string", "null"]
        walletCount:
          type: integer
        solPerWallet:
          type: ["number", "null"]
        totalSolSpent:
          type: ["number", "null"]
        status:
          type: string
        launchType:
          type: string
        performanceUsd:
          type: ["number", "null"]
        performanceSol:
          type: ["number", "null"]
        txHash:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    BundleInput:
      type: object
      required: [ownerAddress, tokenName, tokenSymbol, walletCount]
      properties:
        ownerAddress:
          type: string
        tokenName:
          type: string
        tokenSymbol:
          type: string
        tokenDescription:
          type: string
        tokenImageUrl:
          type: string
        walletCount:
          type: integer
        solPerWallet:
          type: number
        walletIds:
          type: array
          items:
            type: integer
    VampInput:
      type: object
      required: [ownerAddress, sourceTokenAddress, walletCount]
      properties:
        ownerAddress:
          type: string
        sourceTokenAddress:
          type: string
        walletCount:
          type: integer
        solPerWallet:
          type: number
    BundleStats:
      type: object
      required: [totalLaunches, bestLaunchUsd, totalSolSpent, successRate]
      properties:
        totalLaunches:
          type: integer
        bestLaunchUsd:
          type: number
        bestLaunchSol:
          type: number
        totalSolSpent:
          type: number
        successRate:
          type: number
        avgWalletsPerBundle:
          type: number

    # Bump Bots
    BumpBot:
      type: object
      required: [id, ownerAddress, tokenAddress, walletCount, speed, budgetSol, status, createdAt]
      properties:
        id:
          type: integer
        ownerAddress:
          type: string
        tokenAddress:
          type: string
        tokenName:
          type: ["string", "null"]
        walletCount:
          type: integer
        speed:
          type: string
        budgetSol:
          type: number
        solPerWallet:
          type: ["number", "null"]
        bumpSize:
          type: ["number", "null"]
        bumpsPerHour:
          type: ["integer", "null"]
        costPerBump:
          type: ["number", "null"]
        totalBumps:
          type: ["integer", "null"]
        bumpsExecuted:
          type: ["integer", "null"]
        estimatedDurationHours:
          type: ["number", "null"]
        status:
          type: string
        startedAt:
          type: ["string", "null"]
        stoppedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    BotInput:
      type: object
      required: [ownerAddress, tokenAddress, walletCount, speed, budgetSol]
      properties:
        ownerAddress:
          type: string
        tokenAddress:
          type: string
        walletCount:
          type: integer
        speed:
          type: string
        budgetSol:
          type: number
    BotEstimateInput:
      type: object
      required: [walletCount, speed, budgetSol]
      properties:
        walletCount:
          type: integer
        speed:
          type: string
        budgetSol:
          type: number
    BotEstimate:
      type: object
      required: [bumpSize, bumpsPerHour, costPerBump, totalBumps, estimatedDurationHours, solPerWallet]
      properties:
        bumpSize:
          type: number
        bumpsPerHour:
          type: integer
        costPerBump:
          type: number
        totalBumps:
          type: integer
        estimatedDurationHours:
          type: number
        solPerWallet:
          type: number
        isValid:
          type: boolean
        validationMessage:
          type: ["string", "null"]

    # Token
...