# lib/api-zod/src/generated/types/bumpBot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Svarog - Solana Token Launcher Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BumpBot {
  id: number;
  ownerAddress: string;
  tokenAddress: string;
  /** @nullable */
  tokenName?: string | null;
  walletCount: number;
  speed: string;
  budgetSol: number;
  /** @nullable */
  solPerWallet?: number | null;
  /** @nullable */
  bumpSize?: number | null;
  /** @nullable */
  bumpsPerHour?: number | null;
  /** @nullable */
  costPerBump?: number | null;
  /** @nullable */
  totalBumps?: number | null;
  /** @nullable */
  bumpsExecuted?: number | null;
  /** @nullable */
  estimatedDurationHours?: number | null;
  status: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  stoppedAt?: string | null;
  createdAt: string;
  updatedAt?: string;
}